fix(ios): K/N 绑定细节批量修正（cinterop OptIn / NSData.create+writeToFile 扩展 import / zlib UByteVar reinterpret / skiko Paint+FilterQuality / 通知常量与枚举名 / UInt 时钟参数）

+++ shared/src/iosMain/kotlin/com/nainiuzhen/wiki/utils/StartupTime.ios.kt
@@ -38,6 +38,6 @@ actual fun appStartElapsedMs(): Long {
 @OptIn(ExperimentalForeignApi::class)
 private fun monotonicMs(): Long = memScoped {
     val ts = alloc<timespec>()
-    clock_gettime(CLOCK_MONOTONIC, ts.ptr)
+    clock_gettime(CLOCK_MONOTONIC.toUInt(), ts.ptr)
     ts.tv_sec * 1000L + ts.tv_nsec / 1_000_000L
 }
